requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
aiofiles==23.2.1
python-multipart==0.0.9
boto3

